refactor: 핸드폰 번호 10자리인 경우 슬라이스 수정
